Development

+++ src/stores/attendances.ts
@@ -19,14 +19,21 @@ export type Attendance = {
   is_am_out_rectified: boolean
   is_pm_in_rectified: boolean
   is_pm_out_rectified: boolean
-  is_am_leave: boolean
-  is_pm_leave: boolean
+  is_am_absent: boolean
+  is_pm_absent: boolean
   is_overtime_applied: boolean
   overtime_in: string | null
   overtime_out: string | null
   is_overtime_in_rectified: boolean
   is_overtime_out_rectified: boolean
   employee: Employee
+  attendance_images: AttendanceImage[]
+}
+
+export type AttendanceImage = {
+  image_path: string
+  image_type: string
+  created_at: string
 }
 
 export type AttendanceTableFilter = {
@@ -59,7 +66,9 @@ export const useAttendancesStore = defineStore('attendances', () => {
 
     let query = supabase
       .from('attendances')
-      .select('*, employee:employee_id (*)')
+      .select(
+        '*, employee:employee_id (id, firstname, lastname), attendance_images (image_path, image_type, created_at)',
+      )
       .order(column, { ascending: order })
       .range(rangeStart, rangeEnd)
 
@@ -105,6 +114,13 @@ export const useAttendancesStore = defineStore('attendances', () => {
   }
 
   async function deleteAttendance(id: number) {
+    const { data, error: deleteImagesError } = await supabase
+      .from('attendance_images')
+      .delete()
+      .eq('attendance_id', id)
+
+    if (deleteImagesError) return { data, error: deleteImagesError }
+
     return await supabase.from('attendances').delete().eq('id', id).select()
   }
